+++ src/main/java/com/octagram/pollet/survey/domain/model/SurveyPointHistory.java
@@ -1,6 +1,6 @@
 package com.octagram.pollet.survey.domain.model;
 
-import com.octagram.pollet.member.domain.Member;
+import com.octagram.pollet.member.domain.model.Member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
