+++ src/main/java/com/octagram/pollet/survey/domain/model/SurveyResponse.java
@@ -2,7 +2,7 @@ package com.octagram.pollet.survey.domain.model;
 
 import java.time.LocalDateTime;
 
-import com.octagram.pollet.member.domain.Member;
+import com.octagram.pollet.member.domain.model.Member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
